Fix auto sleep setting bugs (#225)
* Fix sleep settings not resetting properly

Clicking the reset button in SettingsPopup calls loadSettings, but
loadSettings calls setAutoStartApplication via setProperty, and
setAutoStartApplication calls saveSettings. This prevents the sleep
settings from resetting properly. I moved the saveSettings call to the
QML to fix this.

* Fix auto sleep enable when sleepAfter isn't changed.

The value of sleepAfter needs to be updated when auto sleep is
enabled, otherwise it will remain zero and the checkbox setting will
not be saved.

Co-authored-by: Nick <nick@i386.io>