shared: Don't check the Cyttsp5 REL bit
The reMarkable vendor kernel incorrectly sets the REL bit for the Cyprus
Cyttsp5 driver. The check in Oxide currently relies on this bug.

The in progress upstream Cyttsp5 kernel driver doesn't set the REL bit.
This patch removes the check from Oxide so that it works on both the
mainline kernel as well as the vendor kernel.

Signed-off-by: Alistair Francis <alistair@alistair23.me>